* Tshì-giām html kah widget thiah--khui
