+++ test/test_html.py
@@ -8,17 +8,13 @@ from django.test.testcases import TestCase
 
 from tuitse import THAU_JI, LIAN_JI, KHIN_SIANN_JI
 from tuitse.widget import KiamTsaNuaUi
+from tuitse.html import tuitse_html
 
 
-class ThongKueTshiGiam(TestCase):
-
-    def _mock_hamsik(self, tinliat):
-        mock = Mock()
-        mock.檢查 = json.dumps(tinliat)
-        return KiamTsaNuaUi().檢查結果(mock)
+class TanGuanTshiGiam(TestCase):
 
     def test_ok(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('好', 'hó', THAU_JI, True),
         ])
         soup = BeautifulSoup(kiatko, "html.parser")
@@ -27,7 +23,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_nng_su(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('真', 'tsin', THAU_JI, True),
             ('好', 'hó', THAU_JI, True),
         ])
@@ -37,7 +33,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_siangji_e_lomaji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('姑', 'koo', THAU_JI, True),
             ('娘', 'niû', LIAN_JI, True),
         ])
@@ -47,7 +43,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_siangji_e_hanji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('姑', 'koo', THAU_JI, True),
             ('娘', 'niû', LIAN_JI, True),
         ])
@@ -57,7 +53,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_khinsiann_hu(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('轉', 'tńg', THAU_JI, True),
             ('去', 'khì', KHIN_SIANN_JI, True),
         ])
@@ -67,7 +63,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_tsuanlo_su(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('koo', 'koo', THAU_JI, True),
             ('niu', 'niu', LIAN_JI, True),
         ])
@@ -77,7 +73,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_khinsiann_e_tsuanlo_su(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('tńg', 'tńg', THAU_JI, True),
             ('khì', 'khì', KHIN_SIANN_JI, True),
         ])
@@ -87,7 +83,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_hanlo_su_e_hanji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('芳', 'phang', THAU_JI, True),
             ('kòng', 'kòng', LIAN_JI, True),
             ('kòng', 'kòng', LIAN_JI, True),
@@ -98,7 +94,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_hanlo_su_e_lomaji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('芳', 'phang', THAU_JI, True),
             ('kòng', 'kòng', LIAN_JI, True),
             ('kòng', 'kòng', LIAN_JI, True),
@@ -109,7 +105,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_kinnkinn叫_e_hanji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('kinn', 'kinn', THAU_JI, True),
             ('kinn', 'kinn', LIAN_JI, True),
             ('叫', '叫', LIAN_JI, True),
@@ -120,7 +116,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_kinnkinn叫_e_lomaji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('kinn', 'kinn', THAU_JI, True),
             ('kinn', 'kinn', LIAN_JI, True),
             ('叫', '叫', LIAN_JI, True),
@@ -131,7 +127,7 @@ class ThongKueTshiGiam(TestCase):
         )
 
     def test_tsuanhan_su(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('帥', '帥', THAU_JI, True),
             ('哥', '哥', LIAN_JI, True),
         ])
@@ -148,21 +144,21 @@ class KingKoTshiGiam(TestCase):
         return KiamTsaNuaUi().檢查結果(mock)
 
     def test_khiam_hanji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('', 'hó', THAU_JI, False),
         ])
         soup = BeautifulSoup(kiatko, "html.parser")
         self.assertEqual(soup.find_all('rt')[0]['class'], ['fail'])
 
     def test_khiam_lomaji(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('好', '', THAU_JI, False),
         ])
         soup = BeautifulSoup(kiatko, "html.parser")
         self.assertEqual(soup.find_all('rb')[0]['class'], ['fail'])
 
     def test_khiam_bue(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('姑', 'koo', THAU_JI, True),
             ('娘', '', LIAN_JI, False),
         ])
@@ -171,7 +167,7 @@ class KingKoTshiGiam(TestCase):
         self.assertEqual(soup.find_all('rb')[2]['class'], ['fail'])
 
     def test_khiam_tiongng(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('歡', 'huann', THAU_JI, True),
             ('歡', '', LIAN_JI, True),
             ('喜', 'hí', LIAN_JI, True),
@@ -184,14 +180,14 @@ class KingKoTshiGiam(TestCase):
         self.assertEqual(lomaji, 'huann  hí')
 
     def test_bo_tuitang(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('好', 'bái', THAU_JI, False),
         ])
         soup = BeautifulSoup(kiatko, "html.parser")
         self.assertEqual(soup.find_all('rb')[0]['class'], ['fail'])
 
     def test_tiongng_bo_tuitang(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('行', 'kiânn', THAU_JI, True),
             ('轉', 'ng', KHIN_SIANN_JI, False),
             ('去', 'i', LIAN_JI, False),
@@ -202,7 +198,7 @@ class KingKoTshiGiam(TestCase):
         self.assertEqual(soup.find_all('rb')[4]['class'], ['fail'])
 
     def test_lianjihu_mai_laklohlai(self):
-        kiatko = self._mock_hamsik([
+        kiatko = tuitse_html([
             ('tánn', '', THAU_JI, False),
             ('pān', '', LIAN_JI, False),
         ])
